fix: restore conversational search and clean up types
For RAG implementations the sources might have urls, but also might be searching an internal datastore like ElasticSearch or Milvus that contains documents that are not accessible on the web. Needed to bring this functionality back (and also clean it up).

Also, we lost passage highlight abilities and brought those back, too.